Criação de um construtor para o AppDbContext e mudança do tipo da variável valor_unitário para double

+++ AvonaleRA/Models/Produto.cs
@@ -5,7 +5,7 @@
 
         public int Id { get; set; }
         public string nome { get; set; }
-        public decimal valor_unitario { get; set; }
+        public double valor_unitario { get; set; }
         public int qtde_estoque { get; set; }
     }
 }
